UPD: BasicProfiler ScopeObject more versatile
git-svn-id: https://svn.rwth-aachen.de/repos/vrgroup-svn/projects/Vista/trunk@23136 33c38b33-09d2-4561-8d4d-1bfa2f58f885

+++ VistaCoreLibs/VistaDeviceDrivers/VistaCHAI3DHapticDevicesDriver/VistaCHAI3DHapticDevicesDriver.cpp
@@ -220,7 +220,7 @@ bool VistaCHAI3DHapticDevicesDriver::Connect()
 
 	// get a handle to the haptic device with this device name
 	// if this is not successful, try to take the first available device
-	for (int i=0;i<numHapticDevices;i++)
+	for ( int i = 0; i < (int)numHapticDevices; i++ )
 	{
 		cHapticDeviceInfo a_deviceSpecification;
 		m_pHapticDeviceHandler->getDeviceSpecifications(a_deviceSpecification, i);
